Split achievements.rs into achievements/ subdirectory
Split crates/simulation/src/achievements.rs (643 lines) into focused modules
under achievements/ for better maintainability:
- types.rs (234 lines): Achievement enum, AchievementReward, AchievementTracker, AchievementNotification
- systems.rs (287 lines): check_achievements system and internal helpers
- tests.rs (106 lines): all unit tests
- mod.rs (25 lines): thin facade with re-exports and plugin registration
